refactoring, move blocking server and client to separate package
