+++ results/1_server_1_clients_blocking.txt
@@ -1,47 +1,47 @@
 ------------------------------
 Iteration: 1
 Number of clients: 1
-Duration nanos: 7185542
+Duration nanos: 8461333
 ------------------------------
 Iteration: 2
 Number of clients: 1
-Duration nanos: 422791
+Duration nanos: 428167
 ------------------------------
 Iteration: 3
 Number of clients: 1
-Duration nanos: 302542
+Duration nanos: 367208
 ------------------------------
 Iteration: 4
 Number of clients: 1
-Duration nanos: 269916
+Duration nanos: 411833
 ------------------------------
 Iteration: 5
 Number of clients: 1
-Duration nanos: 323416
+Duration nanos: 364500
 ------------------------------
 Iteration: 6
 Number of clients: 1
-Duration nanos: 322750
+Duration nanos: 407708
 ------------------------------
 Iteration: 7
 Number of clients: 1
-Duration nanos: 345083
+Duration nanos: 351625
 ------------------------------
 Iteration: 8
 Number of clients: 1
-Duration nanos: 325667
+Duration nanos: 305708
 ------------------------------
 Iteration: 9
 Number of clients: 1
-Duration nanos: 382667
+Duration nanos: 297917
 ------------------------------
 Iteration: 10
 Number of clients: 1
-Duration nanos: 340417
+Duration nanos: 269208
 ------------RESULTS-----------
 Number of clients: 1
 Number of iterations: 10
-Min duration nanos: 269916
-Max duration nanos: 7185542
-Average duration nanos: 1022079.1
-Average duration millis: 1.0220791
+Min duration nanos: 269208
+Max duration nanos: 8461333
+Average duration nanos: 1166520.7
+Average duration millis: 1.1665207
